divide utils module and refactor
